moved parsers into their own subnamespace
