+++ src/Parse/JsonBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Docify
+namespace Docify.Parse
 {
     class JsonBuilder
     {
