+++ src/Parse/MarkdownBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Docify
+namespace Docify.Parse
 {
     class MarkdownBuilder
     {
