+++ src/Parse/ParsedMember.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Docify
+namespace Docify.Parse
 {
     enum ParsedMemberType
     {
